Consolidate proxy url construction
