+++ src/ReverseProxy/Service/RuntimeModel/Transforms/PathStringTransform.cs
@@ -8,7 +8,7 @@ using System.Text;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
 
-namespace ReverseProxy.Core.Service.RuntimeModel.Transforms
+namespace Microsoft.ReverseProxy.Service.RuntimeModel.Transforms
 {
     public class PathStringTransform : RequestParametersTransform
     {
