+++ test/ReverseProxy.Tests/Middleware/DestinationInitializerMiddlewareTests.cs
@@ -51,7 +51,8 @@ namespace Microsoft.ReverseProxy.Middleware.Tests
                 matcherSummary: null,
                 priority: null,
                 backendOrNull: backend1,
-                aspNetCoreEndpoints: aspNetCoreEndpoints.AsReadOnly());
+                aspNetCoreEndpoints: aspNetCoreEndpoints.AsReadOnly(),
+                requestParamterTransforms: null);
             var aspNetCoreEndpoint = CreateAspNetCoreEndpoint(routeConfig);
             aspNetCoreEndpoints.Add(aspNetCoreEndpoint);
             var httpContext = new DefaultHttpContext();
@@ -99,7 +100,8 @@ namespace Microsoft.ReverseProxy.Middleware.Tests
                 matcherSummary: null,
                 priority: null,
                 backendOrNull: backend1,
-                aspNetCoreEndpoints: aspNetCoreEndpoints.AsReadOnly());
+                aspNetCoreEndpoints: aspNetCoreEndpoints.AsReadOnly(),
+                requestParamterTransforms: null);
             var aspNetCoreEndpoint = CreateAspNetCoreEndpoint(routeConfig);
             aspNetCoreEndpoints.Add(aspNetCoreEndpoint);
             var httpContext = new DefaultHttpContext();
