+++ test/ReverseProxy.Tests/Middleware/LoadBalancerMiddlewareTests.cs
@@ -63,7 +63,8 @@ namespace Microsoft.ReverseProxy.Middleware.Tests
                 matcherSummary: null,
                 priority: null,
                 backendOrNull: backend1,
-                aspNetCoreEndpoints: aspNetCoreEndpoints.AsReadOnly());
+                aspNetCoreEndpoints: aspNetCoreEndpoints.AsReadOnly(),
+                requestParamterTransforms: null);
             var aspNetCoreEndpoint = CreateAspNetCoreEndpoint(routeConfig);
             aspNetCoreEndpoints.Add(aspNetCoreEndpoint);
             var httpContext = new DefaultHttpContext();
@@ -116,7 +117,8 @@ namespace Microsoft.ReverseProxy.Middleware.Tests
                 matcherSummary: null,
                 priority: null,
                 backendOrNull: backend1,
-                aspNetCoreEndpoints: aspNetCoreEndpoints.AsReadOnly());
+                aspNetCoreEndpoints: aspNetCoreEndpoints.AsReadOnly(),
+                requestParamterTransforms: null);
             var aspNetCoreEndpoint = CreateAspNetCoreEndpoint(routeConfig);
             aspNetCoreEndpoints.Add(aspNetCoreEndpoint);
             var httpContext = new DefaultHttpContext();
